python/requirements.txt
appdirs==1.4.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
idna==3.3
joblib==1.1.0
numpy==1.23.4
pillow==9.3.0
packaging==21.3
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.9
requests==2.27.1
resampy==0.2.2
scikit-learn==1.1.1
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.9
wincertstore==0.2